scripts/installer/requirements.txt
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

# Copyright (c) 2026 Battelle Energy Alliance, LLC.  All rights reserved.

requests>=2.31
psutil>=5.9
colorama>=0.4.6
distro==1.9.0
python-dotenv>=1.2.1
ruamel.yaml>=0.18.17
customtkinter>=5.2
Pillow>=10.0
pythondialog>=3.5
cairosvg>=2.7